dl_deployment/requirements.txt
azure-common==1.1.26
azure-core==1.9.0
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==11.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.3.0
azureml==0.2.7
azureml-core==1.19.0
azureml-tensorboard==1.19.0
Flask==1.1.2
Flask-RESTful==0.3.8
ndg-httpsclient==0.5.1
networkx==2.5
numpy==1.19.4
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
PyYAML
joblib